feat: create event screen functionality
This commit adds functionality to the create event screen, allowing users to create new events with details such as title, description, website
, address, city, and date/time. It also includes the ability to upload an image for the event.
The commit also includes a default form for new events.
The commit also updates the image service to use the default Firebase storage instance.
The commit also updates the firestore database to store form fields in
 a subcollection of the form document.
The commit also adds a new method to the firestore database to get a list of checkpoints for an event.